Normalize Deployment yamls. (#6432)
This adjusts the layout used by each of our Deployments, so that they can be more easily lined up and compared.  This also inlines the Service definition into each of the Deployments (so that it is more evident when they don't exist!) and adds them where missing (with one notable exception); one of my previous changes discovered that in some mesh configurations (as we run in the smoke test) that Pods exposing ports (e.g. metrics, profiling) can fail to come up if there isn't a Service over them.  The notable exception here is Istio, which is the one controller that we allow to opt-out of the service mesh (it is allowed to be istio-aware because it is an istio component).

This also adjusts DOWN the resource allocations assigned to most of our separable controllers, which had copied the block from our core controller, which reconciles 5-10x more resources than any of them.  I have used the values from our autoscaling controller as a standin, which is 3x smaller.